requirements.txt
pyspark
gcloud
psycopg2-binary
cx-Oracle
mysql-connector-python